requirements.txt
Django==2.0.3

djangorestframework==3.8.2
djangorestframework_jwt==1.11.0
django-cors-headers==2.2.0
django-rest-swagger==2.2.0


django-debug-toolbar==1.9




pyyaml==3.12


tzlocal==1.5.1

pytz==2018.5
